Make openWebBrowser an extension on Context
